# backendnew/docs/swagger.yaml
basePath: /
definitions:
...
  v1.AssetCatalogueResponse:
    properties:
      assets:
        items:
          $ref: '#/definitions/v1.AssetSummary'
        type: array
      total:
        type: integer
    required:
    - assets
    - total
    type: object
...
  v1.AssetDetailsResponse:
    properties:
      asset:
        $ref: '#/definitions/v1.AssetDetails'
    required:
    - asset
    type: object
...
  v1.DiscoverAssetsRequest:
    properties:
      hosts:
        example:
        - example.com
        - 192.168.1.1
        items:
          type: string
        type: array
    required:
    - hosts
    type: object
...
  v1.ErrorResponse:
    properties:
      code:
        type: integer
      details:
        additionalProperties:
          type: string
        type: object
      error:
        type: string
    required:
    - code
    - error
    type: object
...
  v1.StartAllAssetsScanRequest:
    properties:
      asset_types:
        example:
        - domain
        - ip
        - service
        items:
          type: string
        type: array
      scripts:
        items:
          type: string
        type: array
    type: object
...
  v1.StartAssetScanRequest:
    properties:
      scripts:
        example:
        - vulnerability_scan.lua
        - port_scan.lua
        items:
          type: string
        type: array
    type: object
  v1.StartAssetScanResponse:
    properties:
      asset_id:
        type: string
      job_id:
        type: string
      message:
        type: string
      started_at:
        type: string
    required:
    - asset_id
    - job_id
    - message
    - started_at
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@example.com
    name: API Support
    url: http://www.example.com/support
  description: A powerful asset discovery and scanning API with Lua script support
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: Asset Scanner API
  version: "1.0"
paths:
  /assets/{id}:
    get:
      description: Get detailed information about a specific asset including scan
        results
      parameters:
      - description: Asset ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v1.AssetDetailsResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/v1.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v1.ErrorResponse'
      summary: Get asset details
      tags:
      - assets
  /assets/{id}/scan:
    post:
      consumes:
      - application/json
      description: Start scanning a specific asset with specified scripts
      parameters:
      - description: Asset ID
        in: path
        name: id
        required: true
        type: string
      - description: Scan configuration
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/v1.StartAssetScanRequest'
      produces:
      - application/json
      responses:
        "202":
          description: Accepted
          schema:
            $ref: '#/definitions/v1.StartAssetScanResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/v1.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/v1.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v1.ErrorResponse'
      summary: Start asset scan
      tags:
      - assets
  /assets/catalogue:
    get:
      description: Retrieve all discovered assets for 2D view
      parameters:
      - description: Filter by asset type
        in: query
        name: type
        type: string
      - description: Filter by asset status
        in: query
        name: status
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v1.AssetCatalogueResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v1.ErrorResponse'
      summary: Get asset catalogue
      tags:
      - assets
  /assets/discover:
    post:
      consumes:
      - application/json
      description: Start asset discovery for a list of hosts using integrated recon
        service
      parameters:
      - description: Hosts to discover
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/v1.DiscoverAssetsRequest'
      produces:
      - application/json
      responses:
        "202":
          description: Accepted
          schema:
            $ref: '#/definitions/v1.DiscoverAssetsResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/v1.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v1.ErrorResponse'
      summary: Discover assets
      tags:
      - assets
  /assets/scan:
    post:
      consumes:
      - application/json
      description: Start scanning all assets with specified scripts
      parameters:
      - description: Scan configuration
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/v1.StartAllAssetsScanRequest'
      produces:
      - application/json
      responses:
        "202":
          description: Accepted
          schema:
            $ref: '#/definitions/v1.StartAllAssetsScanResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/v1.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v1.ErrorResponse'
      summary: Start scan of all assets
      tags:
      - assets
  /business-units:
    get:
      consumes:
      - application/json
      description: Retrieve all business units
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v1.ListBusinessUnitsResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v1.ErrorResponse'
      summary: List business units
      tags:
      - business-units
    post:
      consumes:
      - application/json
      description: Create a new business unit with the specified name
      parameters:
      - description: Business unit creation request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/v1.CreateBusinessUnitRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/v1.BusinessUnitResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/v1.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/v1.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v1.ErrorResponse'
      summary: Create business unit
      tags:
      - business-units
  /business-units/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a business unit by its ID
      parameters:
      - description: Business unit ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v1.GenericStatusResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/v1.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v1.ErrorResponse'
      summary: Delete business unit
      tags:
      - business-units
    get:
      consumes:
      - application/json
      description: Retrieve a business unit by its ID
      parameters:
      - description: Business unit ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v1.BusinessUnitResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/v1.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v1.ErrorResponse'
      summary: Get business unit
      tags:
      - business-units
    put:
      consumes:
      - application/json
      description: Update a business unit's name
      parameters:
      - description: Business unit ID
        in: path
        name: id
        required: true
        type: string
      - description: Business unit update request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/v1.UpdateBusinessUnitRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v1.BusinessUnitResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/v1.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/v1.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/v1.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v1.ErrorResponse'
      summary: Update business unit
      tags:
      - business-units
...
  /checklist/asset/{id}:
    get:
      consumes:
      - application/json
      description: Retrieve all checklist items applicable to a specific asset with
        their current status
      parameters:
      - description: Asset ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.DerivedChecklistItem'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/v1.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v1.ErrorResponse'
      summary: Get asset-specific checklist items
      tags:
      - checklist
  /checklist/business-unit/{businessUnitId}:
    get:
      consumes:
      - application/json
      description: Retrieve all global checklist items with their current status for
        a specific business unit
      parameters:
      - description: Business Unit ID
        in: path
        name: businessUnitId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.DerivedChecklistItem'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/v1.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/v1.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v1.ErrorResponse'
      summary: Get business unit checklist items
      tags:
      - checklist
  /checklist/business-unit/status:
    post:
      consumes:
      - application/json
      description: Set the status (yes/no/na) of a checklist item for a specific business
        unit
      parameters:
      - description: Business unit checklist status update request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/v1.SetBusinessUnitChecklistStatusRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/v1.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/v1.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v1.ErrorResponse'
      summary: Set business unit checklist item status
      tags:
      - checklist
  /checklist/coverage/summary:
    get:
      consumes:
      - application/json
      description: Get a summary of compliance coverage showing which assets are covered
        by compliance checks
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v1.ErrorResponse'
      summary: Get compliance coverage summary
      tags:
      - checklist
  /checklist/global:
    get:
      consumes:
      - application/json
      description: Retrieve all global checklist items with their current status
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.DerivedChecklistItem'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v1.ErrorResponse'
      summary: Get global checklist items
      tags:
      - checklist
  /checklist/status:
    post:
      consumes:
      - application/json
      description: Set the status (yes/no/na) of a checklist item, either global or
        asset-specific
      parameters:
      - description: Status update request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.SetStatusRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/v1.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v1.ErrorResponse'
      summary: Set checklist item status
      tags:
      - checklist
  /checklist/templates:
    get:
      consumes:
      - application/json
      description: Retrieve all available checklist item templates with covered assets
        that are not compliant (status "no")
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.DerivedChecklistItem'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v1.ErrorResponse'
      summary: List all checklist templates with non-compliant asset coverage
      tags:
      - checklist
  /checklist/templates/upload:
    post:
      consumes:
      - application/json
      description: Upload a JSON file containing checklist templates that will overwrite
        all existing templates
      parameters:
      - description: Templates upload request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.UploadTemplatesRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/v1.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v1.ErrorResponse'
      summary: Upload checklist templates from JSON
      tags:
      - checklist
  /files:
    get:
      description: List all file attachments for a specific checklist key
      parameters:
      - description: Checklist key (e.g., global:item1 or asset:assetId:item1)
        in: query
        name: checklistKey
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.FileAttachmentSummary'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/v1.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v1.ErrorResponse'
      summary: List file attachments
      tags:
      - files
  /files/{fileId}:
    delete:
      description: Delete a file attachment and remove it from MinIO storage
      parameters:
      - description: File ID
        in: path
        name: fileId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/v1.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/v1.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v1.ErrorResponse'
      summary: Delete file attachment
      tags:
      - files
    get:
      description: Get metadata for a file attachment
      parameters:
      - description: File ID
        in: path
        name: fileId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.FileAttachment'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/v1.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/v1.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v1.ErrorResponse'
      summary: Get file information
      tags:
      - files
  /files/{fileId}/download:
    get:
      description: Download a file attachment directly
      parameters:
      - description: File ID
        in: path
        name: fileId
        required: true
        type: string
      produces:
      - application/octet-stream
      responses:
        "200":
          description: File content
          schema:
            type: file
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/v1.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/v1.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v1.ErrorResponse'
      summary: Download file
      tags:
      - files
...
  /files/upload:
    post:
      consumes:
      - multipart/form-data
      description: Upload a file directly as part of a checklist item
      parameters:
      - description: Checklist key (e.g., global:item1)
        in: formData
        name: checklist_key
        required: true
        type: string
      - description: File description
        in: formData
        name: description
        type: string
      - description: File to upload
        in: formData
        name: file
        required: true
        type: file
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/model.FileUploadResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/v1.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v1.ErrorResponse'
      summary: Upload file
      tags:
      - files
...
  /incidents:
    get:
      consumes:
      - application/json
      description: Retrieve all incidents with optional filtering
      parameters:
      - description: Filter by stages (comma-separated)
        in: query
        name: stages
        type: string
      - description: Filter by cause tags (comma-separated)
        in: query
        name: causeTags
        type: string
      - description: Filter by significant incidents only
        in: query
        name: significant
        type: boolean
      - description: Filter by recurring incidents only
        in: query
        name: recurring
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v1.ListIncidentsResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v1.ErrorResponse'
      summary: List incidents
      tags:
      - incidents
    post:
      consumes:
      - application/json
      description: Create a new incident record with initial details
      parameters:
      - description: Incident creation request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/v1.CreateIncidentRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/v1.IncidentResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/v1.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v1.ErrorResponse'
      summary: Create a new incident
      tags:
      - incidents
  /incidents/{id}:
    delete:
      consumes:
      - application/json
      description: Delete an incident by ID
      parameters:
      - description: Incident ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v1.GenericStatusResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/v1.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/v1.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v1.ErrorResponse'
      summary: Delete incident
      tags:
      - incidents
    get:
      consumes:
      - application/json
      description: Retrieve a specific incident by its ID
      parameters:
      - description: Incident ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v1.IncidentResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/v1.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/v1.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v1.ErrorResponse'
      summary: Get incident by ID
      tags:
      - incidents
    put:
      consumes:
      - application/json
      description: Update an existing incident record
      parameters:
      - description: Incident ID
        in: path
        name: id
        required: true
        type: string
      - description: Incident update request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/v1.UpdateIncidentRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v1.IncidentResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/v1.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/v1.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v1.ErrorResponse'
      summary: Update incident
      tags:
      - incidents
  /incidents/stats:
    get:
      consumes:
      - application/json
      description: Retrieve statistics about incidents
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v1.IncidentStatsResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v1.ErrorResponse'
      summary: Get incident statistics
      tags:
      - incidents
  /incidents/summaries:
    get:
      consumes:
      - application/json
      description: Retrieve incident summaries with optional filtering
      parameters:
      - description: Filter by stages (comma-separated)
        in: query
        name: stages
        type: string
      - description: Filter by cause tags (comma-separated)
        in: query
        name: causeTags
        type: string
      - description: Filter by significant incidents only
        in: query
        name: significant
        type: boolean
      - description: Filter by recurring incidents only
        in: query
        name: recurring
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v1.ListIncidentSummariesResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v1.ErrorResponse'
      summary: List incident summaries
      tags:
      - incidents
  /jobs/{id}:
    get:
      description: Get the status and progress of a job
      parameters:
      - description: Job ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v1.JobStatusResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/v1.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v1.ErrorResponse'
      summary: Get job status
      tags:
      - jobs
swagger: "2.0"